Added (non-tested) support for calculating and sending the data arrival rate.

+++ core/EmiPacketHeader.h
@@ -27,8 +27,8 @@ public:
     EmiPacketFlags flags;
     EmiPacketSequenceNumber sequenceNumber; // Set if (flags & EMI_SEQUENCE_NUMBER_PACKET_FLAG)
     EmiPacketSequenceNumber nak; // Set if (flags & EMI_NAK_PACKET_FLAG)
-    uint32_t linkCapacity; // Set if (flags & EMI_LINK_CAPACITY_PACKET_FLAG)
-    uint32_t arrivalRate; // Set if (flags & EMI_ARRIVAL_RATE_PACKET_FLAG)
+    float linkCapacity; // Set if (flags & EMI_LINK_CAPACITY_PACKET_FLAG)
+    float arrivalRate; // Set if (flags & EMI_ARRIVAL_RATE_PACKET_FLAG)
     EmiPacketSequenceNumber rttResponse; // Set if (flags & EMI_RTT_RESPONSE_PACKET_FLAG)
     
     // The maximum value of the RTT response delay is 255 ms.
